# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: MeowHome API
  version: 0.2.0
  description: |
    猫宅 MeowHome 后端 API。移动优先的家庭多猫照护、记录、健康聚合与 AI 辅助。
    所有业务资源均通过 FamilyScope 授权，时间字段统一返回 RFC 3339（UTC 存储）。
servers:
  - url: /api/v1
    description: 默认前缀
...
paths:
...
  /families/{familyId}/agent/messages:
    get:
      tags: [agent]
      summary: 家庭 Agent 消息列表（仅 visibility=family，不含私有聊天）
      description: |
        过滤不可识别的 type/status 时返回 400，不默认为全部数据。
        游标 before 为 `generated_at|id`，不允许跨家庭或绕过权限。
        仅返回 family 可见消息。默认 20 条，最多 50 条；next_cursor 始终存在，
        空字符串表示没有下一页。日期窗口按家庭 IANA 时区的本地日历计算，
        UTC 查询区间使用左闭右开边界。
      operationId: listAgentMessages
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: type
          in: query
          schema:
            type: string
            enum: [patrol_abnormal, patrol_weight, chat_answer, reminder_draft]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, dismissed, expired]
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 家庭消息列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/agent/messages/{messageId}:
    get:
      tags: [agent]
      summary: Agent 消息详情
      operationId: getAgentMessage
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 消息详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /families/{familyId}/agent/messages/{messageId}/dismiss:
    post:
      tags: [agent]
      summary: 忽略 Agent 消息（幂等）
      description: 忽略仅隐藏消息，不撤销已创建的正式提醒。
      operationId: dismissAgentMessage
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 忽略成功
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /families/{familyId}/agent/sessions:
    get:
      tags: [agent]
      summary: 当前用户在本家庭的 Agent 会话列表
      operationId: listAgentSessions
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionListEnvelope'
  /families/{familyId}/agent/sessions/{sessionId}/messages:
    get:
      tags: [agent]
      summary: 本人私有会话消息（游标分页）
      operationId: listAgentSessionMessages
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话消息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionMessagesEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        example: '01J'
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: 无权访问该家庭资源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
        data:
          description: 业务数据
        request_id:
          type: string
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        request_id:
          type: string
...
    AgentMessage:
      type: object
      required: [id, role, visibility, type, severity, title, body, generated_at, model]
      properties:
        id: { type: string }
        session_id: { type: string, description: 对话消息所属会话；巡检消息可空 }
        role:
          type: string
          enum: [user, assistant, tool]
        visibility:
          type: string
          enum: [family, private]
        family_id: { type: string }
        user_id: { type: string, description: 系统消息可空 }
        cat_id: { type: string, description: 单猫消息的明确归属 }
        type:
          type: string
          enum: [patrol_abnormal, patrol_weight, chat_answer, reminder_draft]
        severity:
          type: string
          enum: [info, warning, danger]
        title: { type: string, maxLength: 200 }
        body: { type: string }
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/AgentEvidence'
        action_suggestions:
          type: array
          items:
            $ref: '#/components/schemas/AgentAction'
        draft_version: { type: integer, minimum: 0 }
        draft_reminder:
          $ref: '#/components/schemas/AgentReminderInput'
        draft_expires_at: { type: string, format: date-time }
        action_status:
          type: string
          enum: [pending, confirmed, dismissed, expired]
        display_status:
          type: string
          enum: [dismissed]
          description: 忽略后从家庭消息列表隐藏；已确认提醒的 action_status 保持 confirmed
        generated_at: { type: string, format: date-time }
        model:
          type: string
          description: 规则引擎或实际 Provider 模型名；规则巡检为 rule-engine-v1
        disclaimer: { type: string }
    AgentEvidence:
      type: object
      required: [source_type, source_id]
      description: |
        R-03/R-05 的 source_type=reminder 且 source_id 为提醒 ID；R-04
        引用最近称重记录 ID，若从未称重则以 cat_profile 引用猫咪建档 ID；R-06 使用
        family_scope 和家庭 ID，并在 excerpt 写明查询的本地日期范围。
      properties:
        source_type:
          type: string
          enum: [record, reminder, trend, weight, cat_profile, family_scope]
        source_id: { type: string, description: 记录/提醒 ID；趋势类可空 }
        cat_ids:
          type: array
          items: { type: string }
        occurred_at: { type: string, format: date-time }
        excerpt: { type: string, description: 原始文案节选，禁止改写诊断 }
    AgentAction:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [view_records, view_trend, create_reminder]
        title: { type: string }
        cat_id: { type: string, description: 单猫导航/提醒必填；家庭级留空 }
        days: { type: integer, minimum: 1, maximum: 90 }
        payload:
          type: object
          description: 仅导航类允许携带限定键
    AgentReminderInput:
      type: object
      required: [type, title]
      properties:
        cat_id: { type: string, description: 猫咪 ID 或 both；家庭级留空 }
        type:
          type: string
          enum: [medication, vaccine, deworm, checkup, weight, inventory, water, custom]
        title: { type: string, maxLength: 200 }
        subtitle: { type: string }
        time: { type: string, description: 展示用 time_label 兼容字段 }
        scheduled_at: { type: string, format: date-time, description: 可计算的绝对时间（UTC 存储） }
        timezone: { type: string, description: IANA 时区，展示用 }
        icon: { type: string }
        rule: { type: string }
...
    AgentSessionSummary:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    AgentMessageListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [messages, next_cursor]
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentMessage'
                next_cursor: { type: string }
    AgentMessageEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AgentMessage'
...
    AgentSessionListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [sessions, next_cursor]
              properties:
                sessions:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentSessionSummary'
                next_cursor: { type: string }
    AgentSessionMessagesEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [messages, next_cursor]
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentMessage'
                next_cursor: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []